+++ apps/web/src/components/image.tsx
@@ -3,12 +3,8 @@ import type { ImageProps as UnpicImageProps } from "@unpic/react";
 
 export type ImageProps = UnpicImageProps;
 
-export function Image({
-  alt = "",
-  layout = "constrained",
-  ...props
-}: ImageProps) {
-  return <UnpicImage alt={alt} layout={layout} {...props} />;
+export function Image({ alt = "", ...props }: ImageProps) {
+  return <UnpicImage alt={alt} {...props} />;
 }
 
 export { UnpicImage };